Updated to work with Projectile again
After the Projectile dependency was removed, I'm not sure if it was due to a
change in projectile's handling of things or not, but this stopped working with
my installation. Any time project.el functions are executed they return `nil'
because projectile is being used instead.

Primary change here is to create a new const which evaluates whether or not
Projectile is available to use, if so it loads it. Then it just checks if this
const is true (technically the value is 'projectile') when doing anything that
requires project.el functionality and instead uses the projectile version of
said command instead.

I basically followed along with PR #13 to undo what was done to remove Projectile.