Adds `multi-vterm-project-root` function
`multi-vterm-project-root` checks for a global variable
`multi-vterm-projectile-installed-p` to see if the user has installed
projectile. If that symbol is not bound then we create it and set it to the
result of trying to require projectile. If the user doesn't have projectile in
their load-path, this will return `nil`. Finally we set up our function to
return either the `projectile-project-root` or the `project-root`.

I've refactored this so instead to remove the `defconst` from previous commits
(though this can be undone if the preference is for a flatter presentation of
variables created by the package), as well as using this function in all places
where the project-root is accessed.